fix(init): proper floating tag support (#2468)
This change adds a `pullPolicy` configuration to better support floating
tags for OCI artifacts.  This variable is similar to `forceDownload` but
is intended to bring behavior that is more consistent with other
platforms that deal with image containers.  The possible settings for
`pullPolicy` are:

- `Always`: install-dynamic-plugins.py will compare the image digest in
  the remote registry and download the artifact if it has changed,
regardless if the plugin has been downloaded already.
- `IfNotPresent`: install-dynamic-plugins.py will not compare image
  digests and instead only download the artifact if it is not present
already in the dynamic-plugins-root folder.

These settings have also been applied to the NPM downloading method,
however `Always` will always download the remote artifact with no digest
check.  The existing `forceDownload` option is still functional, however
this setting will take precedence.  Eventually `forceDownload` may be
removed.

This change also fixes an issue where if only the URL is changed for an
OCI artifact (say changing it from 1.0 to 1.1) the artifact is
unintentionally deleted during the cleanup phase of the script.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>
Co-authored-by: Stan Lewis <gashcrumb@gmail.com>